@@ -1,65 +1,7 @@
 {
   "name": "react-router-dom",
-  "version": "5.3.4",
-  "description": "DOM bindings for React Router",
-  "homepage": "https://reactrouter.com/",
-  "repository": {
-    "url": "https://github.com/remix-run/react-router.git",
-    "type": "git",
-    "directory": "packages/react-router-dom"
-  },
-  "license": "MIT",
-  "author": "Remix Software <hello@remix.run>",
-  "files": [
-    "LICENSE",
-    "README.md",
-    "BrowserRouter.js",
-    "HashRouter.js",
-    "Link.js",
-    "MemoryRouter.js",
-    "NavLink.js",
-    "Prompt.js",
-    "Redirect.js",
-    "Route.js",
-    "Router.js",
-    "StaticRouter.js",
-    "Switch.js",
-    "cjs",
-    "es",
-    "esm",
-    "index.js",
-    "generatePath.js",
-    "matchPath.js",
-    "modules/*.js",
-    "modules/utils/*.js",
-    "withRouter.js",
-    "warnAboutDeprecatedCJSRequire.js",
-    "umd"
-  ],
-  "main": "index.js",
-  "module": "esm/react-router-dom.js",
-  "sideEffects": false,
-  "scripts": {
-    "build": "rollup -c",
-    "lint": "eslint modules"
-  },
-  "peerDependencies": {
-    "react": ">=15"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.12.13",
-    "history": "^4.9.0",
-    "loose-envify": "^1.3.1",
-    "prop-types": "^15.6.2",
-    "react-router": "5.3.4",
-    "tiny-invariant": "^1.0.2",
-    "tiny-warning": "^1.0.0"
-  },
-  "browserify": {
-    "transform": [
-      "loose-envify"
-    ]
-  },
+  "version": "6.6.1",
+  "description": "Declarative routing for React web applications",
   "keywords": [
     "react",
     "router",
@@ -67,5 +9,41 @@
     "routing",
     "history",
     "link"
-  ]
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/remix-run/react-router",
+    "directory": "packages/react-router-dom"
+  },
+  "license": "MIT",
+  "author": "Remix Software <hello@remix.run>",
+  "sideEffects": false,
+  "main": "./dist/main.js",
+  "unpkg": "./dist/umd/react-router-dom.production.min.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "dependencies": {
+    "@remix-run/router": "1.2.1",
+    "react-router": "6.6.1"
+  },
+  "devDependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "peerDependencies": {
+    "react": ">=16.8",
+    "react-dom": ">=16.8"
+  },
+  "files": [
+    "dist/",
+    "CHANGELOG.md",
+    "LICENSE.md",
+    "README.md",
+    "server.d.ts",
+    "server.js",
+    "server.mjs"
+  ],
+  "engines": {
+    "node": ">=14"
+  }
 }
