@@ -1,5 +1,5 @@
 import React from 'react';
-import { Switch, Route } from 'react-router-dom';
+import { Routes, Navigate, Route } from 'react-router-dom';
 
 import LandingPage from './landingpage';
 import AboutMe from './about/aboutme';
@@ -7,12 +7,14 @@ import Contact from './contact/contact';
 import Projects from './projects/projects';
 
 const Main = () => (
-    <Switch>
-        <Route exact path="/" component={LandingPage}/>
-        <Route path="/aboutme" component={AboutMe}/>
-        <Route path="/contact" component={Contact}/>
-        <Route path="/projects" component={Projects}/>
-    </Switch>
+    <Routes>
+        <Route exact path="/" element={<LandingPage/>}/>
+        <Route path="/aboutme" element={<AboutMe/>}/>
+        <Route path="/contact" element={<Contact/>}/>
+        <Route path="/projects" element={<Projects/>}/>
+        <Route path="*" element={<Navigate to="/" replace />} />
+        <Route path="/*" element={<Navigate to="/" replace />} />
+    </Routes>
 )
 
 export default Main;
